requirements.txt
crewai
fastapi
uvicorn
requests
pydantic
pyyaml
python-dotenv
rich
pytest
